Adapt webstats to read logs from CollecTor.
In this change we:
 - update metrics-lib to 2.2.0,
 - start downloading and processing logs from CollecTor rather than
   from webstats.torproject.org,
 - change log line counts from int to long,
 - remove webstats tests which are now contained in metrics-lib,
 - update the CollecTor page, and
 - take out the beta notice from the Tor web server logs page.

Implements #25520.

+++ src/main/sql/webstats/init-webstats.sql
@@ -22,7 +22,7 @@ CREATE TABLE requests (
   method METHOD NOT NULL,
   resource_id INTEGER REFERENCES resources (resource_id) NOT NULL,
   response_code SMALLINT NOT NULL,
-  count INTEGER NOT NULL,
+  count BIGINT NOT NULL,
   UNIQUE (file_id, method, resource_id, response_code)
 );
